pyproject.toml
[project]
name = "gpt_speaking_coach"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "realtimestt==0.3.99",
    "python-dotenv==1.0.1",
    "requests"
]